requirements.txt
anthropic>=0.39.0
tavily-python>=0.5.0
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
